re-arrange folder structure to allow readmes to generate correctly
