+++ packages/fields-v2/src/components/Field.tsx
@@ -1,6 +1,22 @@
 import { h, type FunctionalComponent } from '@stencil/core';
 import { ICON_NAMESPACE } from 'icons/namespace';
-import type { FieldProps } from './types';
+
+import type { ValidationMessages } from '../types';
+
+export interface FieldProps {
+    /** The label of the field. */
+    label: string;
+    /** The messages to display under the field. */
+    messages?: ValidationMessages;
+    /** If the field is currently invalid. */
+    invalid?: boolean;
+    /** Inline documentation text. */
+    documentation?: string;
+    /** Inline documentation link. */
+    documentationLink?: string;
+    /** Inline documentation link text. */
+    documentationLinkText?: string;
+}
 
 /** Default layout for a Field. */
 export const Field: FunctionalComponent<FieldProps> = (
