+++ packages/fields-v2/src/components/masked-text/input/masked-text-input.demo.tsx
