+++ packages/fields-v2/src/components/number/input/number-input.css
@@ -1,4 +1,4 @@
-@import '../text.css';
+@import '../../input.css';
 
 .input_wrapper {
     position: relative;
