+++ packages/fields-v2/src/components/text/field/usage/example.md
@@ -1,0 +1,65 @@
+```tsx
+import { createValidatedForm } from '@eventstore-ui/forms';
+
+interface Example {
+    text: string;
+    id: string;
+}
+
+const forms = createValidatedForm<Example>({
+    text: '',
+    id: {
+        initialValue: '',
+        validations: [
+            {
+                validator: (v) => v.length === 12,
+                message: 'Please provide a complete Id',
+            },
+        ],
+    },
+});
+
+const onEnter = () => {
+    forms.submit((data) => {
+        console.log(data);
+    });
+};
+
+export default () => (
+    <>
+        <es-input
+            label={'Text'}
+            placeholder={'Write some text'}
+            onEnter={onEnter}
+            {...forms.connect('text')}
+        />
+        <es-input
+            label={'Account Id'}
+            placeholder={'Account Id'}
+            onEnter={onEnter}
+            mask={{
+                mask: '0000-0000-0000',
+                unmask: true,
+                lazy: false,
+                placeholderChar: '_',
+            }}
+            {...forms.connect('id')}
+        />
+        <es-input
+            disabled
+            label={'Disabled'}
+            placeholder={'This is disabled'}
+            {...forms.connect('text')}
+        />
+    </>
+);
+```
+
+```css
+:host {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+```
