+++ packages/fields-v2/src/components/textarea/field/readme.md
@@ -1,0 +1,89 @@
+# f2-textarea-field
+
+
+
+<!-- Auto Generated Below -->
+
+
+## Overview
+
+A textarea field.
+
+## Usage
+
+### Example
+
+```tsx
+import { createValidatedForm } from '@eventstore-ui/forms';
+
+interface Example {
+    text: string;
+}
+
+const form = createValidatedForm<Example>({
+    text: '',
+});
+
+export default () => (
+    <>
+        <es-textarea
+            label={'Text'}
+            placeholder={'Write some text'}
+            {...form.connect('text')}
+        />
+    </>
+);
+```
+
+```css
+:host {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+```
+
+
+
+## Properties
+
+| Property                   | Attribute                 | Description                              | Type                                 | Default     |
+| -------------------------- | ------------------------- | ---------------------------------------- | ------------------------------------ | ----------- |
+| `disabled`                 | `disabled`                | If the input is disabled.                | `boolean \| undefined`               | `undefined` |
+| `documentation`            | `documentation`           | Inline documentation text.               | `string \| undefined`                | `undefined` |
+| `documentationLink`        | `documentation-link`      | Inline documentation link.               | `string \| undefined`                | `undefined` |
+| `documentationLinkText`    | `documentation-link-text` | Inline documentation link text.          | `string \| undefined`                | `undefined` |
+| `inputProps`               | --                        | Pass props directly to the input.        | `undefined \| { [x: string]: any; }` | `undefined` |
+| `invalid`                  | `invalid`                 | If the field is currently invalid.       | `boolean \| undefined`               | `undefined` |
+| `label` _(required)_       | `label`                   | The label of the field.                  | `string`                             | `undefined` |
+| `messages`                 | --                        | The messages to display under the field. | `ValidationMessages \| undefined`    | `undefined` |
+| `name` _(required)_        | `name`                    | The name of the input.                   | `string`                             | `undefined` |
+| `placeholder` _(required)_ | `placeholder`             | The placeholder for the input.           | `string`                             | `undefined` |
+| `readonly`                 | `readonly`                | If the input is editable.                | `boolean \| undefined`               | `undefined` |
+| `value` _(required)_       | `value`                   | The current value of the field.          | `string`                             | `undefined` |
+
+
+## Events
+
+| Event         | Description                                     | Type                                       |
+| ------------- | ----------------------------------------------- | ------------------------------------------ |
+| `fieldchange` | Emitted when the value of the field is changed. | `CustomEvent<FieldChange<string, string>>` |
+
+
+## Dependencies
+
+### Depends on
+
+- [f2-textarea-input](../input)
+
+### Graph
+```mermaid
+graph TD;
+  f2-textarea-field --> f2-textarea-input
+  style f2-textarea-field fill:#f9f,stroke:#333,stroke-width:4px
+```
+
+----------------------------------------------
+
+
