+++ packages/fields-v2/src/components/textarea/input/textarea-input.css
@@ -1,4 +1,4 @@
-@import '../text.css';
+@import '../../input.css';
 
 .input {
     height: 120px;
