+++ packages/fields-v2/src/components/textarea/input/usage/example.md
@@ -1,0 +1,30 @@
+```tsx
+import { createValidatedForm } from '@eventstore-ui/forms';
+
+interface Example {
+    text: string;
+}
+
+const form = createValidatedForm<Example>({
+    text: '',
+});
+
+export default () => (
+    <>
+        <es-textarea
+            label={'Text'}
+            placeholder={'Write some text'}
+            {...form.connect('text')}
+        />
+    </>
+);
+```
+
+```css
+:host {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+```
